// app/lib/goo/interfaces.ts
export interface Vec2<T> {
    x: T;
    y: T;
}

export interface Vec3<T> {
    x: T;
    y: T;
    z: T;
}
// ...
export type SliceTransform = {
    invert: boolean;
    translate: Vec2<number>;
    scale: Vec2<number>;
    rotate: {
        angle: number; // 0-360
        origin: Vec2<number>;
    };
};

export type PrinterDefinition = {
    name: string;
    type: string;
    resolution: Vec2<number>;
    mirror: Vec2<boolean>;
    platform: Vec3<number>;
    resinProfile: string; // 32b

    grayscale: boolean; // 1b, false - (0x00 ~ 0x01), true - (0x00 ~ 0xFF)
    antialiasing: number; // 2b short int
    gray: number; // 2b short int
    blur: number; // 2b short int
};

export type ImageChannels = 1|2|3|4;
export type ImageDescriptor = string | {
    buffer: Buffer;
    channels: ImageChannels;
};